Tag test which fail with AWS v4 signatures
These tests fail at least on the cloudscale infrastructure, but I
suspect these to fail on any RGW instance. The same tests pass with v2
signatures. So the problem is probably somewhere in the header signing
on the client side or in the signed header verification code on the
server side.