Upgrade requests to at least version 2.18.1
The redirection issue which was the reason to pin requests to a
specific version is fixed. At least the test failure on
s3tests.functional.test_s3.test_post_object_success_redirect_action
no longer occurs with version 2.18.1.

The new version solves the issue the old version has with newer versions
of libssl which disable SSLv3 support.

The currently pinned version of requests is from 2012 (!).